+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- Data Workbench enhancements:
+  - Adds log file display
+  - Add column "Created date" in Exported files and logs tables
+  - Handle refresh event after import, export or delete operations
+
 ## [6.28.0] 2025-02-22
 
 - Data Workbench: Allow to configure all settings related to SFDMU export.json file